run/staticsitegenerator/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
google-cloud-storage==1.42.3
google-api-python-client==2.31.0
